requirements.txt
sqlalchemy
requests
beautifulsoup4
discord.py
elasticsearch